+++ backend/canisters/group/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Implement `c2c_active_proposal_tallies` to be called via composite query ([#8275](https://github.com/open-chat-labs/open-chat/pull/8275))
+
+## [[2.0.1814](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1814-group)] - 2025-07-02
+
 ### Changed
 
 - Updated the `FcmData` interface ([8261](https://github.com/open-chat-labs/open-chat/pull/8261))